Move classes out of newabstraction package
