+++ enigma-server/src/main/java/cuchaz/enigma/network/packet/EntryChangeS2CPacket.java
@@ -5,7 +5,7 @@ import java.io.DataOutput;
 import java.io.IOException;
 
 import cuchaz.enigma.network.ClientPacketHandler;
-import cuchaz.enigma.newabstraction.EntryChange;
+import cuchaz.enigma.translation.mapping.EntryChange;
 
 public class EntryChangeS2CPacket implements Packet<ClientPacketHandler> {
