+++ enigma-swing/src/main/java/cuchaz/enigma/gui/Gui.java
@@ -50,8 +50,8 @@ import cuchaz.enigma.gui.renderer.MessageListCellRenderer;
 import cuchaz.enigma.gui.util.*;
 import cuchaz.enigma.network.Message;
 import cuchaz.enigma.network.packet.MessageC2SPacket;
-import cuchaz.enigma.newabstraction.EntryChange;
 import cuchaz.enigma.source.Token;
+import cuchaz.enigma.translation.mapping.EntryChange;
 import cuchaz.enigma.translation.mapping.EntryRemapper;
 import cuchaz.enigma.translation.representation.entry.ClassEntry;
 import cuchaz.enigma.translation.representation.entry.Entry;
